mmchat/requirements.txt
gradio<=4.40
python-dotenv
bs4
jinja2

# for Azure
azure-cognitiveservices-speech

